Factory pagamento atrasado ou em dua

+++ src/main/java/com/aula42/listaDeProdutos/model/valoresProdutos/CalculoDesconto.java
@@ -1,12 +1,14 @@
 package com.aula42.listaDeProdutos.model.valoresProdutos;
 
+import java.math.BigDecimal;
+
 public class CalculoDesconto {
-    public Double descontoProduto(Double valor, Double porcentagemDesconto){
-        Double resultado = valor * (porcentagemDesconto / 100);
+    public Double descontoProduto(BigDecimal valor, BigDecimal porcentagemDesconto){
+        Double resultado = valor * porcentagemDesconto / 100;
         return resultado;
     }
 
-    public Double valorFinal (Double valor, Double valorDoDesconto){
+    public Double valorFinal (BigDecimal valor, Double valorDoDesconto){
         Double resultadoFinal = valor - valorDoDesconto;
         return resultadoFinal;
     }
